Fixing the controller

+++ sassa-robot-arm/scenarios/scenario1.py
@@ -51,7 +51,8 @@ def scenario1(robot_urdf_path="urdf/sassa-robot/robot.urdf", robot_file_path="ur
 
     sassa.forwardKinematics(q_current)
     pin.updateFramePlacements(sassa.model, sassa.data)
-    viz.display(q_current)
+    if enable_viz:
+        viz.display(q_current)
 
     my_state_machine = StateMahineScenario1(sassa, viz, dt, q0_ref)
 
@@ -206,9 +207,10 @@ def scenario1(robot_urdf_path="urdf/sassa-robot/robot.urdf", robot_file_path="ur
         viz.viewer.gui.setCaptureTransform(motion_file_path, node_list)
 
     # Tests
-    viz.viewer.gui.addCurve("world/pinocchio/curve_0", my_state_machine.trajectory0.getAllPoint(dt), Color.lightBrown)
-    viz.viewer.gui.addCurve("world/pinocchio/curve_1", my_state_machine.trajectory1.getAllPoint(dt), Color.lightBlue)
-    viz.viewer.gui.addCurve("world/pinocchio/curve_2", my_state_machine.trajectory2.getAllPoint(dt), Color.lightYellow)
+    if enable_viz == 1:
+        viz.viewer.gui.addCurve("world/pinocchio/curve_0", my_state_machine.trajectory0.getAllPoint(dt), Color.lightBrown)
+        viz.viewer.gui.addCurve("world/pinocchio/curve_1", my_state_machine.trajectory1.getAllPoint(dt), Color.lightBlue)
+        viz.viewer.gui.addCurve("world/pinocchio/curve_2", my_state_machine.trajectory2.getAllPoint(dt), Color.lightYellow)
 
     # main loop, updating the configuration vector q
     for i in range(int(duration / dt)):
@@ -263,13 +265,14 @@ def scenario1(robot_urdf_path="urdf/sassa-robot/robot.urdf", robot_file_path="ur
                 # wait to have a consitente frame rate
                 time.sleep(tsleep)
 
-    viz.viewer.gui.deleteNode("world/pinocchio/curve_0", True)
-    viz.viewer.gui.deleteNode("world/pinocchio/curve_1", True)
-    viz.viewer.gui.deleteNode("world/pinocchio/curve_2", True)
-    viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm2_sasm_0", True)
-    viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm2_sasm_1", True)
-    viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm2_sasm_2", True)
-    viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm3_sasm_0", True)
+    if enable_viz == 1:
+        viz.viewer.gui.deleteNode("world/pinocchio/curve_0", True)
+        viz.viewer.gui.deleteNode("world/pinocchio/curve_1", True)
+        viz.viewer.gui.deleteNode("world/pinocchio/curve_2", True)
+        viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm2_sasm_0", True)
+        viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm2_sasm_1", True)
+        viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm2_sasm_2", True)
+        viz.viewer.gui.deleteNode("world/pinocchio/visuals/arm3_sasm_0", True)
 
     return log_com, log_goal, log_end_effector
